+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 * The is:haspower search works once again.
 * The is:cosmetic search will now work for Destiny 2.
 * Added is:prophecy search which will return all prophecy weapons from CoO.
+* Added is:ikelos search which will return all ikelos weapons from Warmind.
 
 # 4.60.0 (2018-07-08)